完成10.4 完善技能行为

+++ Src/Lib/Common/Data/SkillDefine.cs
@@ -12,14 +12,14 @@ namespace Common.Data
         public SkillType Type { get; set; }
         public TargetType CastTarget { get; set; }
         public int UnlockLevel { get; set; }
-        public float CastRange { get; set; }
+        public int CastRange { get; set; }
         public float CastTime { get; set; }
         public float CD { get; set; }
         public float MPCost { get; set; }
         public bool Bullet { get; set; }
         public float BulletSpeed { get; set; }
         public string BulletResource { get; set; }
-        public float AOERange { get; set; }
+        public int AOERange { get; set; }
         public string SkillAnim { get; set; }
         /// <summary>
         /// 技能持续时间
